+++ cheetah/accelerator/segment.py
@@ -1,7 +1,7 @@
 from copy import deepcopy
 from functools import reduce
 from pathlib import Path
-from typing import Any, Iterator
+from typing import Any, Iterator, Literal
 
 import matplotlib
 import matplotlib.pyplot as plt
@@ -15,7 +15,11 @@ from cheetah.accelerator.element import Element
 from cheetah.accelerator.marker import Marker
 from cheetah.converters import bmad, elegant, nxtables
 from cheetah.particles import Beam, Species
-from cheetah.utils import UniqueNameGenerator, squash_index_for_unavailable_dims
+from cheetah.utils import (
+    UniqueNameGenerator,
+    merge_element_names,
+    squash_index_for_unavailable_dims,
+)
 
 generate_unique_name = UniqueNameGenerator(prefix="unnamed_element")
 
@@ -148,7 +152,9 @@ class Segment(Element):
             else:
                 flattened_elements.append(element)
 
-        return Segment(elements=flattened_elements, name=self.name)
+        return self.__class__(
+            elements=flattened_elements, name=self.name, sanitize_name=False
+        )
 
     def reversed(self) -> "Segment":
         """
@@ -164,10 +170,10 @@ class Segment(Element):
             )
         )
 
-        return Segment(
+        return self.__class__(
             elements=reversed_elements,
             name=f"{self.name}_reversed",
-            sanitize_name=self.sanitize_name,
+            sanitize_name=False,
         )
 
     def transfer_maps_merged(
@@ -218,7 +224,9 @@ class Segment(Element):
                 )
             )
 
-        return Segment(elements=merged_elements, name=self.name)
+        return self.__class__(
+            elements=merged_elements, name=self.name, sanitize_name=False
+        )
 
     def without_inactive_markers(
         self, except_for: list[str] | None = None
@@ -238,13 +246,14 @@ class Segment(Element):
         if except_for is None:
             except_for = []
 
-        return Segment(
+        return self.__class__(
             elements=[
                 element
                 for element in self.elements
                 if not isinstance(element, Marker) or element.name in except_for
             ],
             name=self.name,
+            sanitize_name=False,
         )
 
     def without_inactive_zero_length_elements(
@@ -264,7 +273,7 @@ class Segment(Element):
         if except_for is None:
             except_for = []
 
-        return Segment(
+        return self.__class__(
             elements=[
                 element
                 for element in self.elements
@@ -273,6 +282,7 @@ class Segment(Element):
                 or element.name in except_for
             ],
             name=self.name,
+            sanitize_name=False,
         )
 
     def inactive_elements_as_drifts(
@@ -292,7 +302,7 @@ class Segment(Element):
         if except_for is None:
             except_for = []
 
-        return Segment(
+        return self.__class__(
             elements=[
                 (
                     element
@@ -304,11 +314,56 @@ class Segment(Element):
                         name=element.name,
                         device=element.length.device,
                         dtype=element.length.dtype,
+                        sanitize_name=False,
                     )
                 )
                 for element in self.elements
             ],
             name=self.name,
+            sanitize_name=False,
+        )
+
+    def with_consecutive_elements_merged(
+        self, except_for: list[str] | None = None
+    ) -> "Segment":
+        """
+        Return a segment where consecutive elements of the same type that can be merged
+        are combined into single elements.
+
+        :param except_for: List of names of elements that should not be merged despite
+            being mergeable.
+        :return: Segment with consecutive mergeable elements merged.
+        """
+        if except_for is None:
+            except_for = []
+
+        merged_elements = []
+        current = self.elements[0]
+        for next_element in self.elements[1:]:
+            if current.name not in except_for:
+                if type(current) is Segment:
+                    current = current.with_consecutive_elements_merged(
+                        except_for=except_for
+                    )
+                elif (
+                    type(current) is type(next_element)
+                    and next_element.name not in except_for
+                ):
+                    merged = current.merge(next_element)
+                    if merged is not None:
+                        current = merged
+                        continue  # Don't do merged_elements.append(current) and advance
+
+            merged_elements.append(current)
+            current = next_element
+
+        merged_elements.append(current)
+
+        return self.__class__(
+            elements=merged_elements,
+            name=self.name,
+            sanitize_name=False,
+            metadata=deepcopy(self.metadata),
         )
 
     @classmethod
@@ -501,7 +556,9 @@ class Segment(Element):
                     # If a non-skippable element is found, merge the skippable elements
                     # and append them before the non-skippable element
                     if len(continuous_skippable_elements) > 0:
-                        todos.append(Segment(elements=continuous_skippable_elements))
+                        todos.append(
+                            self.__class__(elements=continuous_skippable_elements)
+                        )
                         continuous_skippable_elements = []
 
                     todos.append(element)
@@ -509,7 +566,7 @@ class Segment(Element):
             # If there are still skippable elements at the end of the segment append
             # them as well
             if len(continuous_skippable_elements) > 0:
-                todos.append(Segment(elements=continuous_skippable_elements))
+                todos.append(self.__class__(elements=continuous_skippable_elements))
 
             for todo in todos:
                 incoming = todo.track(incoming)
@@ -517,10 +574,11 @@ class Segment(Element):
             return incoming
 
     def clone(self) -> "Segment":
-        return Segment(
+        return self.__class__(
             elements=[element.clone() for element in self.elements],
             name=self.name,
             metadata=deepcopy(self.metadata),
+            sanitize_name=False,
         )
 
     def split(self, resolution: torch.Tensor) -> list[Element]:
@@ -529,6 +587,46 @@ class Segment(Element):
             for element in self.elements
             for split_element in element.split(resolution)
         ]
+
+    def merge(self, other: "Segment") -> "Segment | None":
+        return self.__class__(
+            elements=self.elements + other.elements,
+            name=merge_element_names(self.name, other.name),
+            sanitize_name=False,
+            metadata=other.metadata.update(self.metadata),
+        )
+
+    def partition_at(
+        self, element_name: str, mode: Literal["before", "after", "both"] = "both"
+    ) -> tuple[Element, ...]:
+        """
+        Partition the segment into multiple subcells around a named element. If the
+        segment is split neither before nor after the named element, a single-element
+        tuple containing the original segment is returned.
+
+        :param element_name: Name of the element at which the segment is split.
+        :param mode: Mode of partitioning. If "before", the segment is split before the
+            named element. If "after", the segment is split after the named element. If
+            "both", the segment is split both before and after the named element.
+        :return: Segment partition. May contain 1, 2, or 3 subcells depending on whether
+            the segment is split before and/or after the named element.
+        """
+        index = self.element_index(element_name)
+        pre_cell = (
+            self.__class__(self.elements[: index + 1])
+            if mode == "after"
+            else self.__class__(self.elements[:index])
+        )
+        post_cell = (
+            self.__class__(self.elements[index:])
+            if mode == "before"
+            else self.__class__(self.elements[index + 1 :])
+        )
+        return (
+            (pre_cell, self.elements[index], post_cell)  # Two splits: before and after
+            if mode == "both"
+            else (pre_cell, post_cell)  # One split: before or after
+        )
 
     def beam_along_segment_generator(
         self, incoming: Beam, resolution: float | None = None